Replace magic numbers with compile time string length
The ``sizeof()`` macro is evaluated at compile time. The string passed
to it does not end up in the final binary and there is also no counting
happening at runtime.

The ``sizeof()`` operator counts the null terminator too.
So ``sizeof("foo")`` is 4 and not 3.
But this comes in handy because we want to skip over the length of the
command and the slash character in front of it.