Fix assertion when sound fails to be loaded
Conditionally stop the sample in the `UnloadSample` function instead of calling the separate `Stop` function which is not supposed to be used for samples which are not loaded. The `UnloadSample` function also needs to handle samples which are not (fully) loaded to free the sample index when the sound failed to be loaded.

Regression from #9431.