Fix editor file dialog crash when loading from save storage
The value of the `m_FileDialogStorageType` variable changes when navigating to a different storage location, which was causing the file dialog to incorrectly assume that a file is being saved resulting in an assertion error due to `m_pfnFileDialogFunc` being unhandled. This also caused the "Filename" input to be shown instead of the "Search" input.