Fix incorrect loading progress due to 0.7 images
Load the only 0.7 image that exists in `CScoreboard` directly where it's used instead of looping over all 0.7 images and incorrectly incrementing the loading progress.

The incorrect loading progress was especially noticeable when the client is very slow, e.g. on Android emulators and when running with ASan.

See #9151. This does not address that too much unnecessary code is being generated yet.