Add `CMemoryLogger`
Logger that collects messages in memory. This is useful to collect the log messages for a particular operation and show them in a user interface when the operation failed. Messages can also be forwarded to a parent logger if it's set, regardless of this logger's filter, so the messages also appear in the console normally.