Support reconnecting dummy by redirecting to same port
Previously, the redirect message `NETMSG_REDIRECT` would always cause the main connection to be redirected and the dummy to be disconnected no matter on which connection it is received. Now, if the redirect message is received on the dummy connection with the destination port being idential to the current port then the dummy only will be reconnected. Redirecting the dummy connection to a different port is not allowed, as the dummy should not be connected to a different server than the main, as the client assumes that main and dummy use the same map.

Closes #9480.