Overhaul skin loading and refreshing, improvements for 0.7 skins
Skin loading and refreshing
---------------------------

Add `CSkinDescriptor` class to describe the information related to the skin textures of a `CTeeRenderInfo` so the skin textures can be refreshed independently from the other attributes. Add `CManagedTeeRenderInfo` class as wrapper for a `CTeeRenderInfo` and a corresponding `CSkinDescriptor`. Instances of `CManagedTeeRenderInfo` are managed as `std::shared_ptr`s by the gameclient. This allows skin textures of all `CManagedTeeRenderInfo`s to be refreshed dynamically when the underlying skin is refreshed. This replaces the existing `CComponent::OnRefreshSkin` callback function in which all skins had to be refreshed. This allows dynamically loading and unloading individual skins, to support lazy loading and unloading of skins in the future. This fixes skins not being shown for existing chat messages when toggling the `cl_chat_old` setting. This also fixes downloaded skins not being updated for info messages, chat messages and ghosts.

Improve the skin loading by finishing the jobs of downloaded skins at the beginning of each frame instead of whenever skins are being retrieved. This avoids jobs never being finished if a skin currently being loaded stops being used. Furthermore, this allows to limit the number of skins being loaded per frame based on the elapsed frame time to reduce the FPS impact of skin loading.

Avoid refreshing the skin info of active players every time that a player info snap object is processed. Instead, update the skin info of active players after processing all snap objects but only update the skin textures when the skin was changed to avoid many `CSkins::Find` calls happening every frame for every player.

0.7 skin improvements
---------------------

Support refreshing 0.7 skins, skin parts, as well as the xmas hat and bot decoration and refactor the 0.7 skin loading in general. Refreshing 0.7 skins is implemented as described above for 0.6 skins. The `CSkinDescriptor` class describes the textures of both 0.6 and 0.7 skins. Also log the skin part type in error messages as the part name alone is not unique. Closes #8750.

Use the correct skin blood color for the player death effect with 0.7 skins, which was previously the default color white.

Move operator definitions to the source file to avoid the implicit `system.h` dependency in header. Avoid duplicate code by adding `CSkins7::ApplyColorTo` and `CSkins7::CSkinPart::ApplyTo` functions. Extract `CSkins7::LoadSkin` and `CSkins7::LoadSkinPart` functions for readability.

General changes
---------------

Add member variable for client ID and pointer to gameclient to the `CClientData` class. This avoids passing `IsTeamPlay` as a parameter to the `CClientData::UpdateRenderInfo` function and allows adding more functions directly to the `CClientData` class for better modularity.