Fix crash when loading very long localization lines
The client previously crashed when loading a localization file with a replacement line (plus zero terminator) longer than the size of a `CHeap` chunk (65600 bytes) due to the line data being written outside the heap chunk.

This is fixed by allowing `CHeap` to allocate chunks as large as necessary to contain at least one item.

As an alternative the `CHeap` functions could be changed to return `nullptr` if the wanted allocation is too large, which would have to be handled explicitly when loading localization files.