Do not force infinite ammo for ddrace gametypes
As mentioned in #8923 some of the prediction flags are too
generic. If IsDDRace overwrites InfiniteAmmo then the
InfiniteAmmo flag becomes a bit useless and the IsDDrace
a bit hard to understand.

In ddnet-insta I had to turn off DDRace prediction
to fix a ammo glitch with antiping.

https://github.com/ddnet-insta/ddnet-insta/commit/4c51004ffc11847df0c62c22d9ed676ae343a0db

This then comes with other side effects and breaks all the other
predictions that are activated by IsDDRace.

As of right now InfiniteAmmo is set to true by default.
And it is automatically set to false if the snapped player
has some ammo.

https://github.com/ddnet/ddnet/blob/b453784dcafbf95af88b674c9afd7eb0facaf037/src/game/client/gameclient.cpp#L644
https://github.com/ddnet/ddnet/blob/b453784dcafbf95af88b674c9afd7eb0facaf037/src/game/client/gameclient.cpp#L2956-L2957

So the client correctly detects the InfiniteAmmo already on
ddrace and non ddrace servers. Interestingly the InfiniteAmmo
that is used for prediction has nothing todo with the UnlimitedAmmo
GameInfo flag. That seems a bit weird but might not be a problem.