Add translation credits to language settings
Show credits for the selected language file in the language settings.

The credits are retrieved from the respective language file like a regular localization string. The string `"English translation by the DDNet Team"` with context `"Translation credits: Add your own name here when you update translations"` should be translated to the credits for the respective language. All language authors should be listed (without duplicates) in the order in which they first contributed to the language file. Old author comments should be removed from the file and integrated into this string.

Credits for the German translation are provided as an example.

Closes #6949.