Do not rename redirected players to ``(connecting)``
The player is not connecting anymore. It was already connected
and is now being redirected. Could also use the magic string
``(redirected)`` instead.

But I think it is nice to know on drop what name the player used to
have. Because we do have this information already.

This is needed for ddnet-insta to save players round stats
once they get redirected.

https://github.com/ddnet-insta/ddnet-insta/issues/248

Without this fix the stats would be saved on the name ``(connecting)``