Remove ``IsForceBalanced()``
Newer teeworlds also does not send any chat messages from the
gameworld anymore. Which to me sound the wrong place to do
it anyways.

In ddnet there are no teams so they can also never be balanced.
Also ddnet forks can just implement it on tick there is no need
to access the gameworld for it. Also the method was not virtual
so not too interesting for custom controllers.

I intentionally kept ``bool IGameController::CanBeMovedOnBalance(int ClientId)``
because it is still used in modern teeworlds.
And it can be used in ddnet forks to implement team balancing.

This commit does NOT fall under #7777 because it is a cleanup
that benefits ddnet.

The motivation behind this commit is to cleanup gameworld
to make it easier to link in tests.
Which is need for:

https://github.com/ddnet/ddnet/pull/9386#issuecomment-2565473425