Add community name tooltip to password popup, refactoring
Show tooltip with the community name for the community icon in the password popup. Always show the `Name` label instead of only showing it when no community icon is available, which is only the case if the icon could not be loaded.

Cleanup layout code of the password popup to avoid unnecessary layout and make labels adjust their font size for longer localizations instead of overlapping.