Fix 0.7 server info too large with many player and long names
Fixes server crashing when the 0.7 server info is requested containing more than 38 players with names and clans of maximum length.

Only pack as many clients into the 0.7 server info as will fit into a packet. The client count in the server info must be equal to the included number of client entries, so the entire server info is repacked with a reduced number of clients to ensure that 0.7 clients will still accept it. The real number of players is appended to the server name like for the existing `SERVERINFO_VANILLA` when not all clients could be included.

Closes #10064. Supersedes #10065.