Support changing touch button background color
Add properties to the root object of the touch controls configuration to adjust the background color of all touch buttons:

- `"background-color-inactive"`: specifies the background color of inactive touch buttons.
- `"background-color-active"`: specifies the background color of active touch buttons.

For backwards compatibility these properties are allowed to be unset in which case the previous default values will be used.

Colors are specified as hexadecimal strings in the formats `RRGGBBAA`, `RRGGBB`, `RGBA` and `RGB` without any prefix like `#`, e.g. `"A526C440"`. If the alpha value is omitted then it is set to full opacity by default.