Add `--dry-run` optional parameter to `dilate` tool
Add the `--dry-run` optional parameter to check if an image is dilated with the `dilate` tool without modifying the existing file, by copying and comparing the image data in memory.

The `--dry-run` parameter must be the first parameter if used. All following parameters are interpreted as filenames.

With `--dry-run` enabled the tool will exit with error code 0 if all specified files are already dilated.