Rewrite script to check for dilated images in Python
Use `dilate` tool with the added `--dry-run` parameter to check if all `.png` files in the `data` folder are dilated using a Python script instead of converting the original and dilated images to `.bmp` files and comparing their hashes using a Shell script.

This decreases the runtime of the dilate-step in the CI from around 37 seconds to around 6 seconds and makes installing `imagemagick` unnecessary.