Support using `dbg_graphs` without `debug`, refactoring
Render the graphs when `dbg_graphs` is enabled regardless of the `debug` option instead of only rendering them when both options are enabled, to also allow rendering only the graphs.

Move the alternative between the gameclient and editor rendering to the `CClient::Render` function. Move the `IGraphics::Clear` call to the beginning of the `CGameClient::OnRender` function, consistently with the `Clear` call in the `CEditor::OnRender` function.