Distribute tees across all spawn types
Previously if multiple spawn types (default,red,blue) were used.
Only one of them would be considered for spawning.

Now tees are distributed across all spawn points even if different types
are used in the map.

Unless player collision is off then it is still ensured
that race players keep spawning at the same spawn types to avoid
switching bases in fast cap maps on respawn.

This is an improvement because now in maps where collision is
on all spawns are utilizied. Which is the correct behavior for spawns.
Only using one spawn type was a hack which is now removed.
Distributing tees is desired in maps with collision on.
To avoid crowded spawns and smoother gameplay.

It is also less confusing for mappers now to understand which spawn
type will be active when. Previously all spawns could be used.
But not together in one map. Now mappers can use all spawns and mix
them.