Fix overlapping sequence bit packing of 0.6 network chunk header
Previously, the upper two bits of the lower byte of the network chunk sequence were packed into the 0.6 chunk header twice. With `Split` being `4` for 0.6 chunks, the second byte in the chunk header was split between 4 bits of the sequence and 4 bits of the size, but the lower 2 of those sequence bits were the same as the upper 2 bits that are packed into the third chunk header byte. This worked because the same bits are combined with logic-or when unpacking, so it has no effect overall.

For 0.7 chunks with `Split` being `6`, this was changed to pack 2 bits of the sequence and 6 bits of the size, allowing chunks to theoretically be 4x as large (although the maximum packet size would only allow a much smaller increase).

For 0.6 chunks, this is now changed such that only 2 bits of sequence are packed into the second chunk header byte regardless of the `Split` value. The unused bits are now set to `0`, which is compatible with the old approach because it does not change the result of the logic-or when the chunk header is unpacked.