Remove /unpractice command
It seems to be riddled with exploits, and we don't have the time to
clean these up. There are likely more exploits lurking.

I suggest that we get rid of this command until someone steps up to
clean up the messes from the first two bugs of /unpractice, i.e. write
automated tools to figure out which ranks were cheated with it.