Pass more data to the antibot
Currently there is a limited amount of variables you can analyze in your antibot. This PR adds all of the inputs instead of just the aim position.

This also passes the SixUp, DNSBL and authed state to the antibot to be able to make special cases with those.

Reopen of the PR #9096