Teehistorian: Also record non-changing input
Previously, these were skipped.
If multiple inputs were sent on the same tick, this can cause
unreproducible teehistorians:

A tee is falling towards the floor and presses jump.
It doesn't have dj, so will jump as soon as it reaches the floor.

Scenario 1: Upon touching the ground, the tee sends input to not jump.
This new input is recorded by teehistorian, the tee jumps.

Scenario 2: Upon touching the ground, the tee first sends the former
input (to jump) again, followed by an input not to jump.
The first input is not recorded by Teehistorian, as it is the same as
the previous one.
The second input is recorded, as it differs.
However the first input is used for jumping, so the tee will jump.

The teehistorians from scenario 1 and scenario 2 are indistinguishable,
even though the tee only jumps in scenario 2.