Reduce number of threads in job pool, also use in tests
Reduce number of threads in the job pool to `std::max(4, (int)std::thread::hardware_concurrency()) - 2` to reduce the impact of background jobs on gameplay.

Use this number of threads consistently when creating `CEngine` to avoid duplicate code and to consistently test with the same number of threads as all other code.