Fix 0.7 skin offset being based on 0.6 skin data
The `CRenderTools::GetRenderTeeOffsetToRenderedTee` function is used to determine the Y-offset to render skins almost everywhere, but the offset for 0.7 skins was only mostly correct because the 0.6 skin data for the default skin was also set and used for them. Instead, implement the `CRenderTools::GetRenderTeeOffsetToRenderedTee` function for 0.7 skins without using the default 0.6 skin metrics.