Ensure demo playback starts at time 00:00
The first delta time being used in the `CDemoPlayer::Update` function was very large when loading long demos due to last update time being initialized with an old cached value returned by the `CDemoPlayer::Time` function. This was causing demo playback to start later with a delay proportional to the loading time of the demo. Now, demo playback properly starts at the beginning regardless of the loading time.