Support live playback of demos while they are being recorded
Check if more ticks are still being written to the demo 1 second after starting playback to detect a live demo. For live demos, each tick incrementally scan for new tickmarkers and keyframes after the last keyframe to properly update the total time while playing the demo. Detect if the demo stopped being recorded by checking if no more tickmarkers have been written within 2 seconds.

Prevent seeking closer than 2 seconds to the end of live demos to avoid demo playback errors due to incomplete reads of the last demo chunk data which may still be written to by a demo recorder. Also reset the demo player speed to the default when fast-forwarding to the end of a live demo to prevent playback errors.

Show a red/grey dot indicator next to the demo seekbar for live demos. The indicator is red to indicate playback being considered live when viewing the last 3 seconds of the live demo. Clicking the live indicator will seek to the most recent viewable time.

Furthermore, support playing truncated demos again when possible instead of failing to open them entirely. Keyframes will now only be read until the truncation occurs. Skipping to the end of a truncated demo may still end playback with an error due to the chunk data being incomplete. A truncated demo file is indistinguishable from a live demo being read in an incomplete state. The detection for live demos is deactivated after 15 seconds if it keeps failing, to avoid unnecessary scanning of files that are likely truncated and not live demos.