Rewrite and improve settings & commands export script
Rewrite scripts from https://github.com/ddnet/ddnet-scripts/tree/master/client for exporting the HTML tables of the settings, commands and tunings in Python to significantly improve the readability, extensibility and performance.

Fix duplicate table rows for game commands, as they are also registered on the client side for prediction.

Fix constants `SERVER_MAX_CLIENT` and `SERVER_TICK_SPEED` not being replaced with their numeric values anymore.

Fix description strings containing quotes being truncated or entirely missing.

Order all settings, commands and tunings alphabetically instead of arbitrary based on the order in which the files are enumerated.

Add preview for default values of color settings.

Add strict error handling instead of ignoring errors and silently omitting or mangling items.

Move the script to the ddnet repository and add it to the CI to ensure that it stays synchronized with declarations of settings, commands and tunings.