Prevent direct touch when active touch button becomes invisible
Previously, when a command executed by a touch button caused a touch button to become invisible, the direct touch input was incorrectly activated by the still pressed finger of the invisible button. For example, when touching a bind behavior button with command `dummy_disconnect` and visibility `dummy-connected`, the button would become invisible due to the dummy being disconnected, but the direct touch action, e.g. fire, would incorrectly be activated as well. This is prevented by storing the touch fingers of touch buttons that became invisible until the fingers are released.