Fix deadlocks with Emscripten when waiting for threads
With Emscripten, pthreads do not start until the browser's main loop runs again, which causes deadlocks as we assume in our code that threads are started immediately and wait on condition variables being modified in threads directly after starting them. After starting a thread with `pthread_create`, we now call the `emscripten_sleep` function which a duration of 0 milliseconds to briefly return control to the Emscripten main loop to allow the thread to actually start. This is also what SDL does internally after swapping the graphics.