Fix incorrect usage of `glUniform1f` and `glUniform2fv` functions
The OpenGL `primex.vert` shader only uses the `m_LocRotation` and `m_LocCenter` locations when `TW_ROTATIONLESS` is defined in the shader, i.e. when `Rotationless` is `false` in the `InitPrimExProgram` function.

This was causing error messages in the Emscripten client with GL assertions enabled:

```
glUniform1f called with a nonexisting location ID -1!
glUniform2fv called with a nonexisting location ID -1!
```