Add error handling for checking points on first launch
Show error message if the points for the initially chosen player name could not be determined, e.g. because no internet connection is available. Previously, the popup did not indicate errors and incorrectly showed yes and no buttons also while the points are being checked, now a cancel button is shown instead.