Improve moving client window between screens
Support moving the client window between screens using the window manager, e.g. with Windows+Shift+Left/Right on Windows, by handling `SDL_WINDOWEVENT_DISPLAY_CHANGED`. On Windows 10, this works in all fullscreen modes except in windowed mode when in the special state where the window is using the entire screen size without the decoration/border being shown, which e.g. happens when switching to windowed fullscreen and then back to windowed mode.

Do not resize the window to the screen size when changing the screen in windowed mode, because windowed mode is supposed to use the user-defined size of the window and the window manager may resize the window automatically when moving it between screens.

Remove obsolete SDL TODO/workaround for setting the window to bordered when changing the screen, which does not appear to be necessary for any fullscreen mode. Instead, consistently set the borderless flag for all fullscreen modes in both the `CGraphics_Threaded::IssueInit` and `CGraphicsBackend_SDL_GL::SetWindowParams` functions to avoid discrepancy in window behavior between initial launch of client and after changing the fullscreen mode.

Add more error messages to `CGraphicsBackend_SDL_GL::SetWindowScreen` function when the screen could not be changed.

Closes #8191.