Move `CMenus::DoButton_FontIcon` function to `CUi`
This function is also used for other menus and not specific to the demo menus, so it should not be defined in `menus_demo.cpp`.

This is also a prerequisite for other improvements like adding UI support and buttons to the console and for splitting the demo menus into a separate sub-component.