Fix removing open files on Windows by first renaming to temporary
The `DeleteFileW` function only marks the file for deletion but the deletion may not take effect immediately, which can cause subsequent operations using this filename to fail until all handles are closed. The `MoveFileExW` function with the `MOVEFILE_WRITE_THROUGH` flag is guaranteed to wait for the file to be moved on disk, so we first rename the file to be deleted to a random temporary name and then mark that for deletion, to ensure that the filename is usable immediately.

References:

- https://learn.microsoft.com/en-us/windows/win32/api/fileapi/nf-fileapi-deletefilew
- https://learn.microsoft.com/en-us/windows/win32/api/winbase/nf-winbase-movefileexw

Closes #10241.