Fix editor not saving map if temporary failed to be removed
Consider `fs_remove` function to succeed when failing to remove the renamed temporary file, to ensure that the function succeeds if and only if the original file does not exist anymore. When marking a map file as read-only and then saving the map, the read-only file is renamed to `.tobedeleted` but cannot be deleted. This being considered a failure would prevent the editor from renaming the `.tmp` map being currently saved to the original filename, so the user would be left without the map in `.map` format.