Refactor `net_socket_read_wait` function
Remove the function overload accepting the milliseconds as an `int` and instead only use `std::chrono` types and conversions.

Do not allow waiting for sockets indefinitely when time argument is negative, as this functionality is never used and may lead to issues on some systems. Waiting indefinitely should be avoided with the Emscripten client as it blocks the browser's main thread. It should also be avoided on Windows as it may lead to undefined behavior in the presence of asynchronous procedure calls.

Add doxygen documentation.