Improve websocket handling in `net_socket_read_wait` function
Add the `websocket_fd_get` function to check if a websocket's file descriptor is set in a file descriptor set. Previously, only the presence of the largest file descriptor was checked for Websockets. The check was also dependent on the websocket file descriptor being retrieved last and therefore stored in the `sockid` variable.