Work around Windows resource compilation with Ninja and MSYS2
MinGW's version of the Windows resource compiler (windres) fails when used with the CMake generator Ninja, because of incorrect quote handling in windres and the Ninja generator, also influenced by the presence of different shells like `sh`.

The issues are worked around by not defining any strings with quotes in `CMakeLists.txt` anymore. Instead, the definitions `PLATFORM_SUFFIX` and `GAME_RELEASE_VERSION` are stringified in code. To keep existing usages of the `GAME_RELEASE_VERSION` definition the same, the non-string definition is renamed internally to `GAME_RELEASE_VERSION_INTERNAL`.

Closes #9129.