Use `SHGetKnownFolderPath` to determine storage path on Windows
Using the `SHGetKnownFolderPath` function is the intended method of determining to location of the `AppData/Roaming` folder, as opposed to reading the `APPDATA` environment variable.

The `SHGetKnownFolderPath` function is available since Windows Vista, the current compile target.

This also fixes that the client failed to launch when the `APPDATA` environment variable has a trailing slash or backslash, due to the resulting storage path incorrectly containing two consecutive slashes. The path retrieved by the `SHGetKnownFolderPath` function is guaranteed to never have any trailing slash.