Fix incomplete editor grid, refactoring
The editor grid was rendered with a lot more grid lines than required, as the loop used the UI screen width as upper bound (1066 on a 16:9 screen), which is unrelated to the actual number of grid lines that fit on the screen (less than 200). The number of lines is now properly limited based on the size of the mapped screen of the selected group.

Moreover, the grid lines were incorrectly offset and too short in the top-left corner, which was only visible when the UI is hidden. Instead, they were drawn much longer than necessary in the other direction, also based on the unrelated UI screen size.

The grid lines are now also render batched. To avoid changing the line color during rendering of a batch, the gray grid lines are rendered first and the red grid lines afterwards, if the grid factor is larger than 1.