Add checks for maximum number of chunks per packet
Flush the connection when the maximum number of chunks per packet has been reached, which is 255 due to the value being packed into 1 byte in the chunk header. Otherwise, the number of chunks would be truncated when sending a packet with more chunks. This should have been unlikely to happen in practice, but it could be forced to occur by having the server send a lot of small messages like empty snapshots or rcon lines.

Discard any non-control, non-resend packets that specify zero chunks in their header and discard any control packets that do not specify zero chunks in their header.

Add assertions to the `CNetBase::SendPacket` and `CNetServer::SendMsgs` functions to ensure that the number of chunks per packet is valid when the header is packed.