Fix connection-oriented 0.7 packets being unpacked twice
Connection-oriented 0.7 packets were first also incorrectly unpacked as 0.6 packets to determine that they are connection-oriented and subsequently unpacked correctly as 0.7 packets by considering the sixup-state of the current connection. Now, the packet size and flags are unpacked early, so the current connection slot and sixup state for connection-oriented packets can be determined based on the connless-flag before unpacking the remainder of the packet.

Co-authored-by: fokkonaut <35420825+fokkonaut@users.noreply.github.com>