Fix use of uninitialized value when parsing map commands
Instances of `CMapSettingsBackend::CContext` are used without lineinput/UI when parsing commands on map load, but when opening the map editor from the command line, the font size was still uninitialized. Now, the text width calculations are not performed based on the uninitialized font size when the lineinput is not set, i.e. when the context is used outside of rendering.