Add ``CGameServer::OnSetTimedOut`` in preperation for 128 slots
This will come in handy for the planned 128 player support and reduces
the diff of the currently pending pr #9274

This commit alone without 128 player support is a bit weird and empty.
But it is also a rather complicated change. The gameserver,
gamecontroller and antibot get the client drop with a different state
now. This is why reviewing and merging this independently has its
advantages. Reviewing and testing this independently reduces the risk.
And makes reasoning about the 128 player pr easier and merging it safer.