Add documentation index page, documentation in `CONTRIBUTING.md`
Add main page for the doxygen generated documentation in `src/README.md` and link the existing page for the mastersrv. Previously, the documentation for the mastersrv was shown as the main page due to lack of other `README.md` files in the `src` folder.

Use the DDNet icon as the project logo.

Adjust the documentation name, publisher and IDs.

Increase `DOT_GRAPH_MAX_NODES` to fix warnings due to large include graphs.

Write the warnings to a separate file `docs/warn.log` to make it easier to handle them and eventually also check them in the CI.