Improve OpenGL shader compiler/linker error handling
Determine the total length of the shader compiler/linker error output and allocate a buffer that large to fetch the output. In case of the shader compiler, the length was already fetched and passed to the `glGetProgramInfoLog` function, but the underlying buffer had a fixed size, so this could have caused a stack buffer overflow. Shader linker output was previously truncated.