Add separate setting and quick action for quad envelope preview
This editor feature to preview and edit the position envelope points of quads directly is quite useful, but it was previously combined with the `Show info` setting, making it hard to discover. Add a separate setting `Preview quad envelopes` as well as a quick action to toggle the feature independently instead.