Use words instead of numbers for access level commands
The new encouraged way to refer to access levels in configs is
to use the access level name as a string. The magic number
integer is still kept as an backwards compatibility alias.

And there is currently no plan to deprecate the old alias.

Here a example config

```
# old, still works
access_level ban 1

# new, does the same
access_level ban moderator

# old, still works
access_level kick 2

# new, does the same
access_level kick helper
```

This is a preparation for #10681