Add button to start and connect to local server when LAN tab empty
When no LAN servers are running, show a button to start and connect to the local server in the middle of the empty LAN server list.

Refresh the LAN tab after connecting to the server.

Do not try to connect to the server if it failed to be started.

Show confirmation popup when this would disconnect from a server and `cl_confirm_disconnect_time` is set.

Closes #10718. Supersedes #10731.