More descriptive alternative for GetAuthedState
This will reduce the diff and complexity of the planned auth level extension #10681

Even if there will be more auth levels. The auth level admin.
And the auth level "not authed". Will always be special.
So those two get their own method.

This makes the current code easier to read.
And also will come in handy for changing the auth state to a string
because then less code has to be touched.